fixed) 프로세스 진행 관련 버그 수정 1. DB 탐색 데이터 초기화 정리 2. client->c# dll 함수 테스트 버그 수정 3. wrapper Critical Section 추가 : 함수 Call 중 View 창 띄울 시 에러 발생 4. 통신 데이터 전송 테스트 진행 결과 정리

+++ DefectDBManager/KoWeb_BCR/KoWeb/UDP/Packet.h
@@ -236,7 +236,7 @@ public:
 	void	PullAiNgImagePacket(char* buf, int buf_len, int& frameNum, int& totalNg, int& aiClass, float* scores);
 
 	void	PullBcrSearchLotPacket(char* buf, int buf_len, char* lotName, bool& lotNext, int& vendor, bool& useES, bool& useTG, bool& useETC, bool& useBMark);
-	void	MakeAckBcrSearchLotPacket(CString data, long progress);
+	void	MakeAckBcrSearchLotPacket(CString data, double progress);
 	
 	void	PullBcrSearchModelPatcket(char* buf, int buf_len, char* lotName);
 	void	MakeAckBcrSearchModelPacket(CString data, long progress);
